service/requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
pydantic
python-dotenv
alembic
pyjwt
bcrypt